Fix `fakeWeekly()`
It seems winter/summer time shift was producing a bug in the day
substracting function that make the date stay fixed on a single date
after certain date.

Making the hour stay fixed at the beginning of the loop fixed it. We
didn't need specific hours for this, just to pass 24 times to make a
day.

Also did some name refactoring.